Create new Metadata objects with new Graphs
The multiple db.commit calls seem redudant but they're necessary.
This is mainly because the primary key Graph.id doesn't get generated
until commited to the database, and that's needed to create a new
Metadata object, which is in turn needed for the Graph to generate its
short url. Maybe the short_url should be moved to the metadata in the
first place.